+++ Alienation/program.cpp
@@ -28,8 +28,8 @@ int main(int argc, char* argv[])
 {
    bool     bDone = false;  // Bool Variable To Exit Loop
    int      iVirtual_code;
-	unsigned int		dTickCount;	
-	unsigned int		dLastTickCount;
+	unsigned long int		dTickCount;	
+	unsigned long int		dLastTickCount;
    COpenGL	oGame = COpenGL();
 	SDL_Event oEvent;
 	SDL_Joystick *poJoystick;
